Leaderboard Update, Saving and Loading implemented.
For the snowball type it is a placeholder for now same with the time. Players can now save their 3 letter initials from the dropdown menu. Time is float and been formatted to be 00 : 00 : 00. Snowball type is still a string. The leaderboard loads top 10 names based on their completion time of the stage, their times and snowballs are saved as well.